proxy integration tests united
