+++ nifi-nar-bundles/nifi-elasticsearch-bundle/nifi-elasticsearch-processors/src/test/java/org/apache/nifi/processors/elasticsearch/docker/DockerComposeContainerType.java
@@ -2,6 +2,5 @@ package org.apache.nifi.processors.elasticsearch.docker;
 
 public enum DockerComposeContainerType {
     BASIC,
-    PROXY,
-    PROXY_WITH_AUTH
+    PROXY
 }
