+++ nifi-nar-bundles/nifi-elasticsearch-bundle/nifi-elasticsearch-processors/src/test/resources/proxy-squid/squid-auth.conf
@@ -1,15 +1,6 @@
-acl all src 0.0.0.0/0   
-acl localnet src 0.0.0.1-0.255.255.255	
-acl localnet src 10.0.0.0/8		
-acl localnet src 100.64.0.0/10		
-acl localnet src 169.254.0.0/16 	
-acl localnet src 172.16.0.0/12		
-acl localnet src 192.168.0.0/16		
-acl localnet src fc00::/7       	
-acl localnet src fe80::/10  
 acl localhost src 127.0.0.1/32
 acl to_localhost dst 127.0.0.0/8
-acl localnet src 172.18.0.0/16
+acl localnet src 172.19.0.0/16
 
 acl Safe_ports port 443    
 acl Safe_ports port 9200
